requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docker-pycreds==0.4.0
et-xmlfile==1.1.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
iniconfig==1.1.1
numpy==1.22.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
pluggy==1.0.0
promise==2.3
protobuf==4.21.7
psutil==5.9.2
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-mock==3.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.3
PyYAML==6.0
requests==2.28.1
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.12
wandb==0.13.4
